Add automatic value type detection
Previously, if the user didn't specifiy type=array and gave
value=[1,2,3] (or somesuch), the behaviour was very confusing. This was
because the type couldn't account well for arrays being supplied without
being declared.

This commit adds the ability for the type to try and dynamically divine
the intent of the user if they decline to explicitely state what type of
value they are providing. Basically, if they provide an array of more
than one value, the type will be automatically recognized as an array.

Fixes #1